extract character component from party
