+++ src/views/party/PartyComponent.tsx
@@ -1,10 +1,8 @@
 import style from './PartyComponent.module.scss';
-import { MouseEvent, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { CharacterDto, PartySlot } from '../../common/api/.generated';
 import { CharactersClient } from '../../common/api/client';
-import { useGameStore } from '../game/GameStore';
-import { GameView } from '../game/types';
-import { AssetsService } from '../../common/services/AssetsService';
+import { CharacterComponent } from './CharacterComponent';
 
 export const PartyComponent = () => {
     const [party, setParty] = useState<CharacterDto[]>([]);
@@ -75,53 +73,5 @@ export const PartyComponent = () => {
                 ))}
             </div>
         </section>
-    );
-};
-
-const CharacterComponent = ({ character }: { character: CharacterDto }) => {
-    const { setCharacter, setView } = useGameStore();
-
-    const navigateToSkills = (event: MouseEvent) => {
-        event.stopPropagation();
-        setCharacter(character);
-        setView(GameView.Skills);
-    };
-
-    const navigateToEquipment = (event: MouseEvent) => {
-        event.stopPropagation();
-        console.log('TODO - equipment page');
-    };
-
-    return (
-        <>
-            <p className="bold primary">
-                {character.name} | {character.level}
-            </p>
-            <hr />
-            <img
-                src={AssetsService.getCharacterUri(character.imageUri)}
-                alt={character.name}
-                draggable={false}
-            />
-            <hr />
-            <div style={{ display: 'flex', justifyContent: 'center', gap: '1rem' }}>
-                <div className={style.characterNavIcon}>
-                    <img
-                        onClick={(e) => navigateToSkills(e)}
-                        src={AssetsService.getIcon('SKILLS')}
-                        style={{ width: '2rem', height: '2rem' }}
-                        draggable={false}
-                    />
-                </div>
-                <div className={style.characterNavIcon}>
-                    <img
-                        onClick={(e) => navigateToEquipment(e)}
-                        src={AssetsService.getIcon('EQUIPMENT')}
-                        style={{ width: '2rem', height: '2rem' }}
-                        draggable={false}
-                    />
-                </div>
-            </div>
-        </>
     );
 };
